+++ lib/BackgroundJob/DailyClusteringJob.php
@@ -35,8 +35,9 @@ class DailyClusteringJob extends TimedJob {
                 $maxTimeGap = (int)$this->config->getUserValue($uid, 'journeys', 'maxTimeGap', 86400);
                 $maxDistanceKm = (float)$this->config->getUserValue($uid, 'journeys', 'maxDistanceKm', 100.0);
                 $includeGroupFolders = (bool)((int)$this->config->getUserValue($uid, 'journeys', 'includeGroupFolders', 0));
+                $includeSharedImages = (bool)((int)$this->config->getUserValue($uid, 'journeys', 'includeSharedImages', 0));
                 // Home-aware by default; let ClusteringManager detect home automatically per user
-                $result = $this->clusteringManager->clusterForUser($uid, $maxTimeGap, $maxDistanceKm, max(1, $minClusterSize), true, null, null, false, 5, true, $includeGroupFolders);
+                $result = $this->clusteringManager->clusterForUser($uid, $maxTimeGap, $maxDistanceKm, max(1, $minClusterSize), true, null, null, false, 5, true, $includeGroupFolders, $includeSharedImages);
                 if (isset($result['error'])) {
                     $this->logger->warning('Journeys daily job: clustering error', [ 'user' => $uid, 'error' => $result['error'] ]);
                 } else {